+++ validity/CHANGELOG.md
@@ -6,12 +6,12 @@
 
 ### Added
 
-- `prettyValidate`, `validationIsValid`, `prettyValidation`
+* `prettyValidate`, `validationIsValid`, `prettyValidation`
+* `validateNotNaN`, `validateNotInfinite`
 
 ### Changed
 
-- Renamed `prettyValidation` to `prettyValidate` before adding the new `prettyValidation`.
-
+* Renamed `prettyValidation` to `prettyValidate` before adding the new `prettyValidation`.
 * `NaN`, `+Infinity` and `-Infinity` are now considered valid for both `Double` and `Float`.
 
 ## [0.8.0.0] - 2018-08-25